+++ web/components/markdown.vue
@@ -1,8 +1,9 @@
 <template>
   <div class="container">
     <div class="mb-2">
-      <input id="markdown-raw" v-model="render" type="checkbox">
-      <label for="markdown-raw" class="select-none"> Render markdown</label><br>
+      <input id="markdown-raw" v-model="render" type="checkbox" />
+      <label for="markdown-raw" class="select-none"> Render markdown</label
+      ><br />
     </div>
     <!-- eslint-disable-next-line vue/no-v-html -->
     <div v-if="render" class="markdown" v-html="rendered" />
@@ -11,36 +12,38 @@
 </template>
 
 <script>
-import 'prismjs/themes/prism.css'
-import { defineComponent, ref } from 'vue'
-import { marked } from 'marked'
-import prism from 'prismjs'
-import { safeComputed } from '../composables/utils'
+import "prismjs/themes/prism.css";
+import { defineComponent, ref } from "vue";
+import { marked } from "marked";
+import prism from "prismjs";
+import { safeComputed } from "../composables/utils";
 
 export default defineComponent({
   props: {
     value: {
       type: String,
-      default: ''
-    }
+      default: "",
+    },
   },
-  setup (ctx) {
-    const render = ref(true)
-    const rendered = safeComputed(() => marked(ctx.value, {
-      highlight (code, lang) {
-        if (lang === 'ts') {
-          lang = 'js'
-        }
-        const _lang = prism.languages[lang]
-        return _lang ? prism.highlight(code, _lang) : code
-      }
-    }))
+  setup(ctx) {
+    const render = ref(true);
+    const rendered = safeComputed(() =>
+      marked(ctx.value, {
+        highlight(code, lang) {
+          if (lang === "ts") {
+            lang = "js";
+          }
+          const _lang = prism.languages[lang];
+          return _lang ? prism.highlight(code, _lang) : code;
+        },
+      })
+    );
     return {
       render,
-      rendered
-    }
-  }
-})
+      rendered,
+    };
+  },
+});
 </script>
 
 <style scoped>
@@ -51,14 +54,15 @@ export default defineComponent({
 
 <style>
 .markdown {
-  font-family: -apple-system,BlinkMacSystemFont,Segoe UI,Helvetica,Arial,sans-serif,Apple Color Emoji,Segoe UI Emoji !important;
+  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Helvetica, Arial,
+    sans-serif, Apple Color Emoji, Segoe UI Emoji !important;
   color: #24292e !important;
   word-wrap: break-word;
 }
 
 .markdown code {
-  font-family: SFMono-Regular,Consolas,Liberation Mono,Menlo,monospace;
-  padding: .2em .4em;
+  font-family: SFMono-Regular, Consolas, Liberation Mono, Menlo, monospace;
+  padding: 0.2em 0.4em;
   font-size: 85%;
   background-color: #f6f8fa;
   border-radius: 6px;