+++ web/nuxt.config.ts
@@ -1,9 +1,7 @@
-import { defineNuxtConfig } from 'nuxt/config'
+import { defineNuxtConfig } from "nuxt/config";
 
 // https://v3.nuxtjs.org/docs/directory-structure/nuxt.config
 export default defineNuxtConfig({
   ssr: false,
-  modules: [
-    'nuxt-windicss'
-  ]
-})
+  modules: ["nuxt-windicss"],
+});